refactor(REQ-001): widgets/auth 슬라이스 정통 FSD 정리
- auth-section.tsx → AuthSection.tsx (case rename, Linux 빌드 호환)
- AuthSection: default → named export, 화살표 함수
- widgets/auth barrel: 잘못된 AuthPage 별칭 제거, AuthSection named re-export
- pages/riot: deep import → barrel(@/widgets/auth) 경유

동작 변경 없음. 변경 파일 한정 tsc/biome clean, pnpm build 통과.
REQ-001 → specs/done/, 검증 체크리스트·회고 보고서 추가.
