+++ requirements/reports/checklists/REQ-001.md
@@ -1,0 +1,43 @@
+---
+id: REQ-001
+title: widgets/auth 슬라이스 정통 FSD 정리
+mode: refactor
+validated: 2026-05-10
+---
+
+## Checklist
+
+| # | 요구사항 | 구현 위치 | 상태 |
+|---|----------|-----------|------|
+| 1 | `auth-section.tsx` → `AuthSection.tsx` rename (case 정합) | `src/widgets/auth/ui/AuthSection.tsx` (git mv -f, 히스토리 보존) | ✅ |
+| 2 | default export → `export const AuthSection` | `AuthSection.tsx:7` `export const AuthSection = () => { ... };` | ✅ |
+| 3 | `widgets/auth/index.ts`에서 `AuthSection` named re-export | `src/widgets/auth/index.ts` `export { AuthSection } from "./ui/AuthSection";` | ✅ |
+| 4 | `pages/riot/index.tsx`가 `@/widgets/auth` barrel 경유 import | `src/pages/riot/index.tsx:4` `import { AuthSection } from "@/widgets/auth";` | ✅ |
+
+## Acceptance Criteria
+
+| # | 기준 | 검증 방법 | 상태 |
+|---|------|-----------|------|
+| AC1 | `auth-section`을 deep import하는 코드 0건 | `grep -rn 'auth-section' src/` → 0 hit | ✅ |
+| AC2 | `pnpm ci` + `pnpm build` 통과 | 빌드 exit 0. biome ci/tsc 사전 존재 에러는 본 REQ 변경 파일과 무관 (`AuthSection.tsx`/`widgets/auth/index.ts`/`pages/riot/index.tsx` 모두 clean) | ✅ |
+| AC3 | `/riot/` 진입 시 기존과 동일 동작 | 수동 스모크 필요 (LogoButton·LoginButton·자동 로그인 체크박스 렌더 확인) | ⏳ |
+
+## Refactor Checklist (사후 평가)
+
+| # | 항목 | 평가 | 비고 |
+|---|------|------|------|
+| 0 | 사전 결정 기준 | 5/5 Yes | plan과 동일 |
+| 0 | 자동 보류 신호 | none | 작업 시작 시점 빌드 green, import 영향 1곳 |
+| 1 | FSD 준수 | pass | barrel 경유 + 레이어 의존 방향 정합 |
+| 2 | 가독성 / 클린 코드 | pass | PascalCase 파일명, named export, 화살표 함수 |
+| 3 | 명확한 함수 | pass | 함수 본문/책임 변경 없음 |
+| 4 | 모듈화 | pass | barrel public API로 단일화, deep import 차단 |
+| 5 | 성능 | N/A | 동작·렌더 경로 무변경 |
+
+## Out-of-Scope (참고)
+
+같은 슬라이스(`widgets/auth`) 내 잔존 deep import는 본 REQ 범위 밖. 별도 REQ 후보:
+- `pages/sign-up/terms.tsx` → `@/widgets/auth/ui/sign-up-section` (default export, kebab-case)
+- `pages/riot/callback.tsx` → `@/widgets/auth/ui/auth-callback` (default export, kebab-case, tsc 사전 에러 존재)
+
+코드베이스 전반의 사전 존재 위반(entities → features cross-layer, generated 직접 참조 등)도 본 REQ 범위 밖이며 별도 REQ로 다뤄져야 한다.
