+++ requirements/reports/retrospects/REQ-001.md
@@ -1,0 +1,60 @@
+---
+id: REQ-001
+title: widgets/auth 슬라이스 정통 FSD 정리
+mode: refactor
+retrospected: 2026-05-10
+---
+
+## Retrospective Report — 2026-05-10 #1
+
+### ✅ What Went Well
+
+- **case-only rename 안전 처리**: macOS case-insensitive APFS 환경에서 `git mv -f`로 한 번에 처리되어 임시 파일 단계 없이 히스토리를 보존하며 rename 완료. Linux Docker 빌드(case-sensitive) 잠재 깨짐 해소.
+- **변경 영향 범위가 plan과 정확히 일치**: 영향 import 사이트 1곳(`pages/riot/index.tsx`)만 갱신. 인접 코드를 건드리지 않아 PR 노이즈 제로.
+- **변경 파일 한정 검증 모두 clean**: tsc/biome/build 통과. 코드베이스 전반의 사전 존재 위반은 손대지 않음(리팩토링 범위 원칙 준수).
+- **barrel 경유 정상화**: `AuthPage`라는 잘못된 별칭이 제거되고 `AuthSection`이 슬라이스 public API로 일관됨. deep import 진입점 1개 차단.
+
+### ⚠️ What Needs Improvement
+
+- **`Checkbox.onCheckedChange`의 `checked as boolean` 좁힘**: `src/widgets/auth/ui/AuthSection.tsx:34` — Radix `CheckedState`는 `boolean | "indeterminate"`인데 강제 캐스팅. 본 REQ 범위 밖(동작 변경 금지)이라 그대로 유지했으나, indeterminate를 사용하지 않는 의도라면 `checked === true`로 명시하는 것이 안전.
+- **같은 슬라이스 내 잔존 비표준 컴포넌트**: `auth-callback.tsx`, `sign-up-section.tsx` — 둘 다 default export + kebab-case 파일명. 외부에서 deep import 중. 본 REQ가 다룬 동일 종류의 위반이며 별도 REQ로 정리해야 일관성이 완성됨.
+- **`pages/riot/index.tsx`의 `useAuth` 위치**: `@/shared/model/use-auth`에서 import — 정통 위치는 `@/entities/auth/model/useAuth`. 본 REQ 범위 밖이지만 인접 비표준이며 후속 REQ 후보.
+
+### 📋 Action Items
+
+- [ ] **REQ 후보**: `widgets/auth/ui/auth-callback.tsx` → `AuthCallback.tsx` + named export + barrel 경유 (REQ-001과 동형 작업) — 우선순위: medium
+- [ ] **REQ 후보**: `widgets/auth/ui/sign-up-section.tsx` → `SignUpSection.tsx` + named export + barrel 경유 — 우선순위: medium
+- [ ] **REQ 후보**: `shared/model/use-auth.tsx` → `entities/auth/model/useAuth.ts` 이전 (`shared/model/` 비표준 segment 정리의 일환) — 우선순위: low
+- [ ] **수동 스모크**: `pnpm dev` → http://localhost:3000/riot/ 진입하여 LogoButton·LoginButton·자동 로그인 체크박스 렌더 확인 — 우선순위: high (커밋 전)
+
+### 📝 Notes (실행 가능 작업 아님)
+
+- `Checkbox` 캐스팅 개선은 본 슬라이스가 아니라 `shared/ui/checkbox/Checkbox` 호출 패턴 전반의 문제. 일괄 정리는 별도 검토 필요.
+- 코드베이스 전반의 cross-layer/cross-slice 위반(entities → features 다수, generated 직접 참조 등)은 본 REQ 회고 범위를 크게 초과 — 별도 마스터 REQ로 다뤄야 함.
+
+### 📊 Refactor Checklist Post-Evaluation
+
+| # | 항목 | plan 시점 | 현재 | 변동 |
+|---|------|-----------|------|------|
+| 1 | FSD 준수 | pass | pass | = |
+| 2 | 가독성 / 클린 코드 | pass | pass | = |
+| 3 | 명확한 함수 | pass | pass | = |
+| 4 | 모듈화 | pass | pass | = |
+| 5 | 성능 (계측 기반) | N/A | N/A | = |
+| 0 | 동작 동일성 | — | code clean / 수동 스모크 pending | — |
+
+악화된 항목 없음.
+
+### 📊 Quality Score
+
+| 항목 | 점수 (1-5) |
+|------|-----------|
+| FSD 준수 | 5/5 |
+| 타입 안전성 | 4/5 |
+| 코드 완성도 | 5/5 |
+| 에러 핸들링 | N/A |
+| 가독성 / 클린 코드 | 5/5 |
+| 모듈화 | 5/5 |
+| 성능 (계측 기반) | N/A |
+
+타입 안전성이 5가 아닌 이유: 컴포넌트 자체에 새 타입 이슈는 없으나 기존 `checked as boolean` 캐스팅을 그대로 유지(범위 밖). REQ-001 범위 한정으로는 만점에 가깝다.
