+++ src/pages/riot/index.tsx
@@ -1,7 +1,7 @@
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
 import { useEffect } from "react";
 import { useAuth } from "@/shared/model/use-auth";
-import AuthSection from "@/widgets/auth/ui/auth-section";
+import { AuthSection } from "@/widgets/auth";
 
 function AuthPage() {
 	const { isAuthenticated } = useAuth();
